nebula/requirements.txt
anyio==4.0.0
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
ndg-httpsclient==0.5.1
packaging==23.2
Pillow==10.4.0
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
python-multipart==0.0.17
requests==2.31.0
rfc3986==2.0.0
shortuuid==1.0.11
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
starlette==0.41.1
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
